chore: added metadata, maxwidthwrapper and opengraph image to awareness page
